added user verification using the save new login button and + button
